+++ src/core/moves/admixture/AdmixtureEdgeAddCladeResiduals.h
@@ -33,7 +33,7 @@ namespace RevBayesCore {
         // Basic utility functions
         AdmixtureEdgeAddCladeResiduals*              clone(void) const;                                                                  //!< Clone object
         void                            swapNode(DagNode *oldN, DagNode *newN);
-        bool                            isActive(int g) const;
+        bool                            isActive(unsigned long g) const;
         
     protected:
         const std::string&              getMoveName(void) const;                                                            //!< Get the name of the move for summary printing
@@ -53,15 +53,15 @@ namespace RevBayesCore {
         StochasticNode<int>*            admixtureCount;
         DeterministicNode<std::vector<double> >* residuals;
         //ConstantNode<int>*              delayTimer;
-        int                             activeGen;
+        unsigned long                   activeGen;
         bool                            changed;
         bool                            failed;
         bool                            allowSisterAdmixture;
         int                             delay;
-        int                             maxEvents;
-        int                             numEvents;
-        int                             numNodes;
-        int                             numTips;
+        size_t                          maxEvents;
+        size_t                          numEvents;
+        size_t                          numNodes;
+        size_t                          numTips;
         
         // stored objects to undo proposal
         AdmixtureNode*                  storedAdmixtureParent;
